Config: Updated DB host to handle ipv6
Can be set via the square bracket format.
For #5464

+++ tests/TestCase.php
@@ -118,7 +118,7 @@ abstract class TestCase extends BaseTestCase
      * Database config is juggled so the value can be restored when
      * parallel testing are used, where multiple databases exist.
      */
-    protected function runWithEnv(string $name, $value, callable $callback)
+    protected function runWithEnv(string $name, $value, callable $callback, bool $handleDatabase = true)
     {
         Env::disablePutenv();
         $originalVal = $_SERVER[$name] ?? null;
@@ -132,13 +132,17 @@ abstract class TestCase extends BaseTestCase
         $database = config('database.connections.mysql_testing.database');
         $this->refreshApplication();
 
-        DB::purge();
-        config()->set('database.connections.mysql_testing.database', $database);
-        DB::beginTransaction();
+        if ($handleDatabase) {
+            DB::purge();
+            config()->set('database.connections.mysql_testing.database', $database);
+            DB::beginTransaction();
+        }
 
         $callback();
 
-        DB::rollBack();
+        if ($handleDatabase) {
+            DB::rollBack();
+        }
 
         if (is_null($originalVal)) {
             unset($_SERVER[$name]);
